Consolidate code in APS2Pattern
Try to combine again drivers for APS2 and TDM. The compiler functions are separate. Some instructions (like Invalidate and LoadCmpVRam) work differently for APS2 and TDM. Explicit calls to LoadVRam can only be made on the APS2 for now.
